[Cleanup] Remove MegatronTrainer (#576)
